Refactor SSE client transport
